requirements.txt
alembic==1.13.1
astroid==3.0.3
asttokens==2.4.1
blinker==1.7.0
cachelib==0.12.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dill==0.3.8
dnspython==2.6.1
email-validator==2.1.0.post1
executing==2.0.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.48.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
ipython==8.21.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
kiwisolver==1.4.5
lint==1.2.1
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
parso==0.8.3
patsy==0.5.6
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
pylint==3.0.3
pymannkendall==1.4.3
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-memcached==1.62
pytz==2024.1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.27
stack-data==0.6.3
statsmodels==0.14.1
tomlkit==0.12.3
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
WTForms-JSON==0.3.5